Fix comparison of ABSL_MIN_LOG_LEVEL with absl::LogSeverity.
Explicitly cast absl::LogSeverity::kFatal to int for comparison with ABSL_MIN_LOG_LEVEL.

#codehealth

PiperOrigin-RevId: 943916309

+++ src/base/container/flat_set_test.cc
@@ -84,7 +84,8 @@ TEST(FlatSetDeathTest, DuplicateEntries) {
   // Runtime construction with duplicate entries hits LOG(FATAL). Compile-time
   // construction with duplicate entries fails the build instead.
 #ifdef ABSL_MIN_LOG_LEVEL
-  constexpr bool kIsLogFatal = ABSL_MIN_LOG_LEVEL <= absl::LogSeverity::kFatal;
+  constexpr bool kIsLogFatal =
+      ABSL_MIN_LOG_LEVEL <= static_cast<int>(absl::LogSeverity::kFatal);
 #else  // ABSL_MIN_LOG_LEVEL
   constexpr bool kIsLogFatal = true;
 #endif  // ABSL_MIN_LOG_LEVEL
